client/requirements.txt
aio-pika==9.1.2
aiormq==6.7.6
grpcio==1.54.2
grpcio-tools==1.54.2
idna==3.4
multidict==6.0.4
pamqp==3.2.1
pika==1.3.2
protobuf==4.23.2
yarl==1.9.2
